+++ tests/redirect-url.test.ts
@@ -2,12 +2,6 @@ import { FormTypes } from '@oneblink/types'
 import replaceInjectablesWithSubmissionValues from '../src/services/replaceInjectablesWithSubmissionValues'
 import { FormSubmissionResult } from '../src/submission-service'
 
-jest.mock('nanoid/non-secure', () => {
-  return {
-    customAlphabet: () => () => 'abCD5678',
-  }
-})
-
 describe('Form redirect URL', () => {
   const definition: FormTypes.Form = {
     id: 1,